fix: split impls for modules
