+++ Cargo.toml
@@ -12,10 +12,8 @@ libc = "*"
 bedrock = { git = "https://github.com/Pctg-x8/bedrock", features = ["Presentation", "Implements"] }
 num = "*"
 
-peridot-vertex-processing-pack = { path = "vertex-processing-pack" }
 peridot-archive = { path = "archive" }
 peridot-math = { path = "math" }
-peridot-vg = { path = "vg" }
 
 image = "*"
