+++ src/asset.rs
@@ -10,33 +10,28 @@ pub trait PlatformAssetLoader {
     fn get(&self, path: &str, ext: &str) -> IOResult<Self::Asset>;
     fn get_streaming(&self, path: &str, ext: &str) -> IOResult<Self::StreamingAsset>;
 }
-pub trait LogicalAssetData: Sized {
+pub trait LogicalAssetData: Sized
+{
     const EXT: &'static str;
 }
-pub trait FromAsset: LogicalAssetData {
+pub trait FromAsset: LogicalAssetData
+{
     type Error: From<IOError>;
     fn from_asset<Asset: Read + Seek + 'static>(asset: Asset) -> Result<Self, Self::Error>;
     
-    fn from_archive(reader: &mut archive::ArchiveRead, path: &str) -> Result<Self, Self::Error> {
-        let bin = reader.read_bin(path)?;
-        match bin {
+    fn from_archive(reader: &mut archive::ArchiveRead, path: &str) -> Result<Self, Self::Error>
+    {
+        match reader.read_bin(path)?
+        {
             None => Err(IOError::new(ErrorKind::NotFound, "No Entry in primary asset package").into()),
             Some(b) => Self::from_asset(Cursor::new(b))
         }
     }
 }
-pub trait FromStreamingAsset: LogicalAssetData {
+pub trait FromStreamingAsset: LogicalAssetData
+{
     type Error: From<IOError>;
     fn from_asset<Asset: Read + 'static>(asset: Asset) -> Result<Self, Self::Error>;
-}
-use vertex_processing_pack::*;
-impl LogicalAssetData for PvpContainer { const EXT: &'static str = "pvp"; }
-impl FromAsset for PvpContainer {
-    type Error = IOError;
-
-    fn from_asset<Asset: Read + Seek + 'static>(asset: Asset) -> IOResult<Self> {
-        PvpContainerReader::new(BufReader::new(asset)).and_then(PvpContainerReader::into_container)
-    }
 }
 
 use image::{ImageDecoder, ImageResult, ImageError};
