+++ src/lib.rs
@@ -7,9 +7,7 @@ extern crate libc;
 extern crate pathfinder_partitioner;
 extern crate bedrock;
 pub extern crate peridot_math as math;
-pub extern crate peridot_vertex_processing_pack as vertex_processing_pack;
 pub extern crate peridot_archive as archive;
-pub extern crate peridot_vg as vg;
 
 use bedrock as br; use bedrock::traits::*;
 use std::ops::Deref;
@@ -461,42 +459,6 @@ impl RenderPassTemplates {
         b.add_dependency(SubpassDependencyTemplates::to_color_attachment_in(None, 0, true));
 
         return b;
-    }
-}
-
-use std::ffi::CString;
-use vertex_processing_pack::PvpContainer;
-pub struct PvpShaderModules<'d> {
-    bindings: Vec<br::vk::VkVertexInputBindingDescription>, attributes: Vec<br::vk::VkVertexInputAttributeDescription>,
-    vertex: br::ShaderModule, fragment: Option<br::ShaderModule>,
-    vertex_spec_constants: Option<(Vec<br::vk::VkSpecializationMapEntry>, br::DynamicDataCell<'d>)>,
-    fragment_spec_constants: Option<(Vec<br::vk::VkSpecializationMapEntry>, br::DynamicDataCell<'d>)>,
-}
-impl<'d> PvpShaderModules<'d> {
-    pub fn new(device: &br::Device, container: PvpContainer) -> br::Result<Self> {
-        Ok(PvpShaderModules {
-            vertex: br::ShaderModule::from_memory(device, &container.vertex_shader)?,
-            fragment: if let Some(b) = container.fragment_shader {
-                Some(br::ShaderModule::from_memory(device, &b)?)
-            }
-            else { None },
-            bindings: container.vertex_bindings, attributes: container.vertex_attributes,
-            vertex_spec_constants: None, fragment_spec_constants: None
-        })
-    }
-    pub fn generate_vps(&'d self, primitive_topo: br::vk::VkPrimitiveTopology) -> br::VertexProcessingStages<'d> {
-        let mut r = br::VertexProcessingStages::new(br::PipelineShader
-        {
-            module: &self.vertex, entry_name: CString::new("main").expect("unreachable"),
-            specinfo: self.vertex_spec_constants.clone()
-        }, &self.bindings, &self.attributes, primitive_topo);
-        if let Some(ref f) = self.fragment {
-            r.fragment_shader(br::PipelineShader {
-                module: f, entry_name: CString::new("main").expect("unreachable"),
-                specinfo: self.fragment_spec_constants.clone()
-            });
-        }
-        return r;
     }
 }
