+++ vertex-processing-pack/Cargo.toml
@@ -5,4 +5,5 @@ authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
 
 [dependencies]
 bedrock = { git = "https://github.com/Pctg-x8/bedrock" }
+peridot = { path = ".." }
 peridot-serialization-utils = { path = "../serialization-utils" }
