+++ vg/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 authors = ["S.Percentage <Syn.Tri.Naga@gmail.com>"]
 
 [dependencies]
-peridot-math = { path = "../math" }
+peridot = { path = ".." }
 pathfinder_partitioner = { git = "https://github.com/Pctg-x8/pathfinder" }
 # 0.3.1のコミットに合わせる
 font-kit = { git = "https://github.com/pcwalton/font-kit", rev = "e54e9d9546f59dab0ae7d91db052908773488edd" }
